requirements.txt
APScheduler==3.10.4
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
